updating to version 1.0

+++ src/main/java/org/damageprofiler/calculations/LengthDistribution.java
@@ -75,10 +75,10 @@ public class LengthDistribution {
         }
 
         if(!length_distribution_map.containsKey(record_length)){
-            length_distribution_map.put((double)record_length, 1.0);
+            length_distribution_map.put(record_length, 1.0);
         } else {
             double count = length_distribution_map.get(record_length);
-            length_distribution_map.put((double)record_length, count + 1);
+            length_distribution_map.put(record_length, count + 1);
         }
     }
 
@@ -98,9 +98,7 @@ public class LengthDistribution {
         return length_forward;
     }
 
-    public List<Double> getLength_reverse() {
-        return length_reverse;
-    }
+    public List<Double> getLength_reverse() { return length_reverse; }
 
     public List<Double> getLength_all() {
         return length_all;
